+++ daemon/main/DiskService.cpp
@@ -59,7 +59,7 @@ void DiskService::ServiceWork()
 
 void DiskService::CheckDiskSpace()
 {
-	long long freeSpace = Util::FreeDiskSize(g_Options->GetDestDir());
+	int64 freeSpace = Util::FreeDiskSize(g_Options->GetDestDir());
 	if (freeSpace > -1 && freeSpace / 1024 / 1024 < g_Options->GetDiskSpace())
 	{
 		warn("Low disk space on %s. Pausing download", g_Options->GetDestDir());
